♻️ [refactor]: #452 Phase 7: lexer_cross_api.rs を分離
横断（panic 不在 / 不変条件 / 結合）の 4 テストを
`src/lexer/tests/lexer_cross_api.rs` に分離。
テスト本体は Token/Primitive を直接参照しないため `use super::super::Lexer;` のみ。

- 293 件グリーン維持
- 短縮テスト関数名集合不変 / 重複 0
- lexer_cross_api.rs: 92 行 (< 800)

Refs #452
